Message model created

+++ app/Http/Controllers/TaskController.php
@@ -156,6 +156,13 @@ class TaskController extends Controller
      */
     public function destroy(Task $task)
     {
+        if ($task->user_id !== $this->user->id) {
+            return new JsonResponse(
+                'task ' . $task->title . ' cannot be deleted because you are not the owner of it',
+                Response::HTTP_INTERNAL_SERVER_ERROR
+            );
+        }
+
         if ($task->delete()) {
             return new JsonResponse("Task " . $task->name . " was deleted successfully", Response::HTTP_OK);
         }
